ДЗ 16 (без 5 задачи). Задачу 5 (наблюдатель паттерн) я не сделал пока - постараюсь завтра. Остальные задачи выполнил. 1. Полям классов из предметной области переопределить геттеры и сеттеры. Сделал: добавил сеттер с проверкой корректности цифр и геттер с увеличением цены на процент НДС. 2. Создать класс каталог содержащий коллекцию объектов. Класс каталог должен быть Generic. Сделал. Класс CatalogClass.kt 3. Добавить в main activity текстовое поле в которое необходимо выводить содержимое каталога. Сделал: вьюшка textViewCatalogContent 4а. Кнопка add должна брать первый элемент из каталога и добавлять его в singleton. Сделал: функция addNewDrink. Отошел от задания, добавил удаление первого элемента из класса CatalogClass.kt 4б. Убрать кнопку next и previous. Сделал. Можно еще delete убрать - все равно больше не работает. 4с. В текстовое поле из ДЗ 15 выводить весь список из singleton. Сделал: функция printAllInfoFromSingleton
