--- app/src/main/java/by/evgen/an11pazniakhomework14/homework15/Beer.kt
@@ -1,4 +1,0 @@
-package by.evgen.an11pazniakhomework14.homework15
-
-class Beer {
-}
