--- app/src/main/java/by/evgen/an11pazniakhomework14/homework15/Util.kt
@@ -1,7 +1,0 @@
-package by.evgen.an11pazniakhomework14.homework15
-
-fun WhiskeyCola.getInfo() = "name: " + this.name + ", capacity: " +  this.volume + ", alcohol: " + this.alco
-fun Whiskey.getInfo() = "name: " + this.name + ", capacity: " +  this.volume + ", alcohol: " + this.alco
-fun Jorsh.getInfo() = "name: " + this.name + ", capacity: " +  this.volume + ", alcohol: " + this.alco
-fun Vodka.getInfo() = "name: " + this.name + ", capacity: " +  this.volume + ", alcohol: " + this.alco
-fun WhiteRussian.getInfo() = "name: " + this.name + ", capacity: " +  this.volume + ", alcohol: " + this.alco
