+++ app/src/main/java/by/evgen/an11pazniakhomework14/homework15/data/Beer.kt
@@ -1,0 +1,4 @@
+package by.evgen.an11pazniakhomework14.homework15.data
+
+class Beer {
+}
