+++ app/src/main/java/by/evgen/an11pazniakhomework14/homework15/data/Spirits.kt
@@ -1,4 +1,4 @@
-package by.evgen.an11pazniakhomework14.homework15
+package by.evgen.an11pazniakhomework14.homework15.data
 
 abstract class Spirits(open val name: String, open val alco: Int, open val volume: Int) {
     open fun drinkAlone(): String{
