+++ app/src/main/java/by/evgen/an11pazniakhomework14/homework15/data/Whiskey.kt
@@ -1,12 +1,24 @@
-package by.evgen.an11pazniakhomework14.homework15
+package by.evgen.an11pazniakhomework14.homework15.data
+
+import by.evgen.an11pazniakhomework14.homework15.util.tax
 
 open class Whiskey(
     override val name: String = "Whiskey",
-    override val alco: Int = 40,
-    override val volume: Int = 500,
+    override val alco: Int = 43,
+    override val volume: Int = 50,
     val ice: Ice = Ice()
 ) : //агрегация - мы в конструкторе требуем передать нам чужой объект "ice"
     Spirits(name, alco, volume) {
+    open var price: Int = 20
+        set(price) {
+            if ((price > 20) && (price < 200)) {
+                field = price
+            }
+        }
+        get() {
+            return (field * tax).toInt()
+        }
+
     override fun drinkAlone(): String {
         return "You can drink $name alone, but company is better"
     }
