+++ app/src/main/java/by/evgen/an11pazniakhomework14/homework15/util/OrderList.kt
@@ -1,4 +1,6 @@
-package by.evgen.an11pazniakhomework14.homework15
+package by.evgen.an11pazniakhomework14.homework15.util
+
+import by.evgen.an11pazniakhomework14.homework15.data.*
 
 object OrderList {
     var listOfDrinks: ArrayList<Spirits> =
