+++ app/src/main/java/by/evgen/an11pazniakhomework14/old/CatDataClass.kt
@@ -1,4 +1,4 @@
-package by.evgen.an11pazniakhomework14
+package by.evgen.an11pazniakhomework14.old
 
 data class CatDataClass(
     val catAge: Int,
