+++ app/src/main/java/by/evgen/an11pazniakhomework14/old/ExtendDataClassFun.kt
@@ -1,4 +1,4 @@
-package by.evgen.an11pazniakhomework14
+package by.evgen.an11pazniakhomework14.old
 
 fun CatDataClass.countFields(): Int {
     return this.catAge + this.tailLength + this.countPaws
