+++ app/src/main/java/by/evgen/an11pazniakhomework14/old/InheritCatNoDataClass.kt
@@ -1,4 +1,4 @@
-package by.evgen.an11pazniakhomework14
+package by.evgen.an11pazniakhomework14.old
 
 class InheritCatNoDataClass : CatNoDataClass(5, 15) { //Создать класс наследник от data класса
 }
